webmessaging/postMessage-triggered-by-user-activation.html

<!DOCTYPE html>
<title>postMessage's task triggered by user activation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/multipage/web-messaging.html#dom-window-postmessage">
<link rel="help" href="https://html.spec.whatwg.org/multipage/comms.html#dom-messageport-postmessage">
<link rel="help" href="https://html.spec.whatwg.org/multipage/web-messaging.html#dom-broadcastchannel-postmessage">
<link rel="help" href="https://html.spec.whatwg.org/multipage/interaction.html#triggered-by-user-activation">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsers.html#the-rules-for-choosing-a-browsing-context-given-a-browsing-context-name">
<body>
<script>
"use strict";

function canOpenPopup() {
  let popup = window.open("about:blank", "_blank");
  if (popup) {
    popup.close();
    return true;
  }
  return false;
}

function appendAndClickButton(buttonText, onClickCallback) {
    let button = document.createElement("button");
    button.textContent = buttonText;
    button.onclick = onClickCallback;
    document.body.appendChild(button);
    test_driver.click(button);
}

async_test(t => {
  window.onmessage = t.step_func_done(e => {
    assert_equals(e.data, "click");
    assert_true(canOpenPopup(), "onmessage be able to open popup");
  });

  appendAndClickButton("Window's postMessage", t.step_func(() => {
    window.postMessage("click", "*");
  }));
}, "With window");

async_test(t => {
  assert_true("MessageChannel" in self, "The browser must support MessageChannel");

  const channel = new MessageChannel();

  channel.port2.onmessage = t.step_func_done(e => {
    assert_equals(e.data, "click");
    assert_true(canOpenPopup(), "onmessage be able to open popup");
  });

  appendAndClickButton("MessagePort's postMessage", t.step_func(() => {
    channel.port1.postMessage("click");
  }));
}, "With a MessageChannel and its MessagePorts");

async_test(t => {
  assert_true("BroadcastChannel" in self, "The browser must support BroadcastChannel");

  const channel = new BroadcastChannel("channel name");

  channel.onmessage = t.step_func_done(e => {
    assert_equals(e.data, "click");
    assert_true(canOpenPopup(), "onmessage be able to open popup");
  });

  appendAndClickButton("BroadcastChannel's postMessage", t.step_func(() => {
    let myChannel = new BroadcastChannel("channel name");
    myChannel.postMessage("click");
  }));

}, "With a BroadcastChannel");

</script>
</body>
